+++ app/build.gradle
@@ -23,6 +23,7 @@ android {
     testOptions {
         unitTests.returnDefaultValues = true
     }
+    sourceSets { main { res.srcDirs = ['src/main/res', 'src/main/res/raw'] } }
 }
 
 dependencies {
