+++ app/src/main/java/bmtreuherz/artour/Activities/Preferences.kt
@@ -2,23 +2,27 @@ package bmtreuherz.artour.Activities
 
 import android.content.Context
 import android.content.SharedPreferences
+import bmtreuherz.artour.Utilities.BeaconEventBroadcaster.context
 
 /**
  * Used to set all user preferences with setters and getters
  */
 
 class Preferences (context: Context) {
-    val PREFS_FILENAME = "artour.prefs"
-    val LANGUAGE = "language"
-    val prefs: SharedPreferences = context.getSharedPreferences(PREFS_FILENAME, Context.MODE_PRIVATE)
+    companion object {
 
-    fun getLang() : String {
-        return prefs.getString(LANGUAGE, "english")
-    }
+        val PREFS_FILENAME = "artour.prefs"
+        val LANGUAGE = "language"
+        val prefs: SharedPreferences = context.getSharedPreferences(PREFS_FILENAME, Context.MODE_PRIVATE)
 
-    fun setLang(lang:String) {
-        val editor = prefs.edit()
-        editor.putString(LANGUAGE, lang)
-        editor.apply()
+        fun getLang(): String {
+            return prefs.getString(LANGUAGE, "english")
+        }
+
+        fun setLang(lang: String) {
+            val editor = prefs.edit()
+            editor.putString(LANGUAGE, lang)
+            editor.apply()
+        }
     }
 }
