+++ app/src/main/res/values/strings.xml
@@ -31,6 +31,4 @@
     <string name="spanish">Spanish</string>
     <string name="chinese">Chinese</string>
 
-    <string name="back">Back</string>
-
 </resources>
